renamed some variables: output_channel -> component_channel --makes it more clear that the output_channel is also an identifier for the component SIC_COMMON_LOG -> SIC_CLIENT_LOG _stop_event -> _signal_to_stop --signal_to_stop is more accurate, it cues the Component to stop, after which "_stopped" is set
made it so all logger instances are instantiated with a Redis object.
optional argument "client_logger" used to determine whether log messages will be printed to the terminal and written to file (subscribes the client logger to the client's redis channel)

added a COMPONENT_STOP_TIMEOUT variable, which dictates how long to wait for the component to stop safely. A component is considered safely stopped once the "_stopped" event is set.

CM only registers atexit handler if it is the main thread (its own process, as in all cases except when it's a Desktop CM). This is to avoid both the sic_application and the CM doing a shutdown process.